% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cctox_model}
\alias{cctox_model}
\title{Colorectal cancer toxicity model}
\format{
An object of class \code{pois} of length 2.
}
\usage{
cctox_model
}
\description{
A fitted POIS model to a real colorectal cancer toxicity dataset
}
\keyword{datasets}
