% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fit_pois_glmnet_mmdcv}
\alias{fit_pois_glmnet_mmdcv}
\title{Fit POIS-glment with MMD cross-validation}
\usage{
fit_pois_glmnet_mmdcv(
  z,
  r = max(z),
  train_ratio = 0.7,
  nlam = 12,
  lambda = 10^seq(-4, -1.3, length.out = nlam),
  nreps = 2,
  agg_func = mean,
  alpha = 1,
  use_parallel = F,
  ncores = 7,
  symmetrize = TRUE
)
}
\arguments{
\item{z}{is a potentially sparse data array of dimensions: (sample size) x
(data dimension)}

\item{r}{maximum number of levels (K)'}

\item{train_ratio}{train/validation split ratio will be
train_ratio/(1-train_ratio).}

\item{nlam}{number of regularization parameters (lambda) to use; ignored if
lambda is provided.}

\item{lambda}{the vector of regularization parameters to use.}

\item{nreps}{the number of CV splits to average over.}

\item{agg_func}{the aggregation function for the MMDs.}
}
\value{
The lambda vector, the regularization curve, a list of fitted POIS
  models, the index of the optimal model and the optimal lambda
}
\description{
The function fits POIS-glment and applies cross-validation (CV) to choose an
optimal regularization parameter.
}
\details{
The maximum mean discrepency (MMD) is used as the evaluation metric for CV.
The MMD is computed between a sample from the original data and one from the
fitted model. The MMD is computed for a sequence of Gaussian kernels with
varying bandwidths, and aggregated using a user-supplied function.
}
\examples{
out = fit_pois_glmnet_mmdcv(amazon, lambda = 10^seq(-4,-1.3, length.out = 5))
}
