% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{pois}
\alias{pois}
\title{Constructor a POIS model}
\usage{
pois(Theta, Gamma)
}
\arguments{
\item{Theta}{The Theta parameter of POIS model (d x K)}

\item{Gamma}{The Gamma (interaction) parameter of the model (d x d)}
}
\description{
Creates an object of the POIS model class
}
