% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simcctox}
\alias{simcctox}
\title{Colorectal cancer toxicity -- simulated data}
\format{
An object of class \code{dgCMatrix} with 5000 rows and 36 columns.
}
\usage{
simcctox
}
\description{
Simulated data generated from a fitted POIS model (\code{\link{cctox_model}}) to a real colorectal cancer
toxicity dataset.
}
\keyword{datasets}
