% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fit_pois_glmnet_nocv}
\alias{fit_pois_glmnet_nocv}
\title{Fit POIS-glment}
\usage{
fit_pois_glmnet_nocv(
  z,
  r = max(z),
  nlam = 12,
  lambda = 10^seq(-4, -1.3, length.out = nlam),
  alpha = 1,
  use_parallel = F,
  ncores = 7,
  symmetrize = TRUE
)
}
\arguments{
\item{z}{is a potentially sparse count array of dimensions: (sample size) x
(data dimension). The entries of z should take values in {0,1,2,...,r}.}

\item{r}{maximum number of levels (K). Defaults to r = max(z).}

\item{nlam}{number of regularization parameters (lambda) to use; ignored if
lambda is provided.}

\item{lambda}{the vector of regularization parameters to use}
}
\value{
A list of fitted POIS models.
}
\description{
The function fits POIS-glment and returns the entire regularization path. No
cross-validation is used.
}
\examples{
out = fit_pois_glmnet_nocv(amazon, lambda = 10^seq(-4,-1.3, length.out = 5))
}
